+++ com.ibm.wala.cast.js/source/com/ibm/wala/cast/js/ipa/callgraph/ArgumentSpecialization.java
@@ -306,7 +306,7 @@ public class ArgumentSpecialization {
             @Override
             protected String composeEntityName(WalkContext parent, CAstEntity f) {
               if (f == codeBodyEntity) {
-                return super.composeEntityName(parent, f) + '_' + v.getValue().intValue();
+                return super.composeEntityName(parent, f) + '_' + v.getValue();
               } else {
                 return super.composeEntityName(parent, f);
               }
