+++ com.ibm.wala.core.tests/src/com/ibm/wala/core/tests/basic/PrimitivesTest.java
@@ -695,7 +695,7 @@ public class PrimitivesTest extends WalaTestCase {
     // path should be 8, 6, 4, 2, 0
     System.err.println("Path is " + p);
     for (int i = 0; i < p.size(); i++) {
-      Assert.assertTrue((p.get(i)).intValue() == new int[] {8, 6, 4, 2, 0}[i]);
+      Assert.assertTrue(p.get(i) == new int[] {8, 6, 4, 2, 0}[i]);
     }
   }
