+++ com.ibm.wala.core/src/com/ibm/wala/ipa/callgraph/propagation/cfa/ZeroXInstanceKeys.java
@@ -203,7 +203,7 @@ public class ZeroXInstanceKeys implements InstanceKeyFactory {
       HashSet<IClass> smushees = HashSetFactory.make(5);
       for (Map.Entry<IClass, Integer> e : count.entrySet()) {
         Integer i = e.getValue();
-        if (i.intValue() > SMUSH_LIMIT) {
+        if (i > SMUSH_LIMIT) {
           smushees.add(e.getKey());
         }
       }
@@ -223,7 +223,7 @@ public class ZeroXInstanceKeys implements InstanceKeyFactory {
         if (old == null) {
           count.put(alloc, 1);
         } else {
-          count.put(alloc, old.intValue() + 1);
+          count.put(alloc, old + 1);
         }
       }
     }
