+++ com.ibm.wala.util/src/com/ibm/wala/util/graph/traverse/DFS.java
@@ -131,7 +131,7 @@ public class DFS {
       Integer t1 = order.get(o1);
       Integer t2 = order.get(o2);
       // throws an exception if either node has not been ordered
-      return (t1.intValue() - t2.intValue());
+      return (t1 - t2);
     }
   }
